+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Most recent version is listed first.
 # v0.0.92
 - Update docs on id generation: https://github.com/komuw/ong/pull/420
 - Update dependencies: https://github.com/komuw/ong/pull/428
+- ong/sync: concat errors: https://github.com/komuw/ong/pull/429
 
 # v0.0.91
 - Upgrade to Go v1.22: https://github.com/komuw/ong/pull/408